Fix race condition when the same project is resolved concurrently
When using `--parallel`, it's possible that 2 different configurations being resolved have dependencies
on the same local components. In that case, they may be resolved concurrently, and a race condition
could occur where the lazy initializing `getConsumableConfigurationsHavingAttributes` method would be
called concurrently. This commit fixes this by adding `synchronized` on the method.

It's worth noting that this is not a great fix: it would be better to have a finer grained lock that
just prevents the same local component from being resolved concurrently, because we also defeat the
in-memory caches by doing this.

See: gradle/performance#502