Make configureByMap less expensive
A lot of objects in Gradle's DSL are created by Map
notation, e.g.

```
  compile (group: 'foo', name: 'bar') {
    exclude group: 'baz'
  }
```

Some of these keys are explicitly queried by the
MapNotationConverter for these types, which is already
reasonably efficient thanks to caching. But we actually
support configuring any other property of the object from
the map. Almost no one makes use of that though.

It turned out that we were wrapping the created object in a
BeanDynamicObject in any case, even if no additional properties
were given. Just checking whether the additional properties are
empty removed a lot of memory pressure.