Fix potential deadlock on parallel test execution
Could happen with certain combinations of maxWorkers/maxParallelForks
and forkEvery. One of our performance tests started to exhibit the
deadlock in a reproducible manner in a feature branch.

This commits fix the potential deadlock caused by a bad handling of
forkEvery wrt. worker leases management in ForkingTestClassProcessor.
The worker lease management responsibility is moved to a dedicated
TestClassProcessor, simpler to understand.