Allow nested usage of BuildOperationProcessor
Use a dynamically sized thread pool in BuildOperationProcessor to allow
nested usage. Maximum in-flight work is managed by worker leases, the
size of that thread pool has nothing to do with that. It being caped
at maxWorkers made it impossible to use BuildOperationProcessor inside
a work enqueued to BuildOperationProcessor up to a certain degree. This
commit fixes that.