Avoid intermediate objects for occasional failures
This commit optimizes the `ArtifactBackedResolvedVariant` by avoiding the creation of intermediate maps, for
failures that would normally not happen. This should fix a performance regression introduced by the implementation
of parallel download of artifacts. Failures are occasional, so the map is only created on demand, and an internal
interface is used for coordination.