Renamed `canPerformPreemptiveDownload` -> `shouldPerformPreemptiveDownload`
When true, this flag indicates that an `ArtifactVisitor` is very likely to
require all of the artifact files, and thus will benefit from downloading
all of the artifacts up-front (and in parallel).

When false, this flag indicates that an `ArtifactVisitor` does not require
the downloaded artifact files, so pre-emptive download will be wasteful.