Make script compilation scenario more realistic
Using `GradleVersion` without an import is a performance pitfall.
The class is in the pseudo-internal package "org.gradle.util", for
which there exists no efficient simple name to qualified name mapping,
but which is still in the "default import" list. The result is that we
look up the type `GradleVersion` in all of our default import packages,
leading to thousands of class not found exceptions per reference.

We should deprecate simple name references to org.gradle.util types,
so that users don't run into this problem.