Only kill existing daemons on clean
Killing them before every task is overly pessimistic, as
leftover daemons is mostly a CI problem when running tests
against old, badly behaving Gradle versions. It is almost
never necessary locally. Since we run `clean` on CI anyway,
it is safe to just run the daemon cleanup as part of the `clean`
task.