Extracted `FileSystemSnapshotter` service out of `AbstractFileCollectionSnapshotter`.
The `FileSystemSnapshotter` takes care of snapshotting the current state of a file or directory in the file system and the associated in-memory mirroring.

The `FileCollectionSnapshotter` implementations then take care of transforming these snapshots to arrive at the snapshot for a given file collection when that file collection is used for a particular purpose.