Use a semaphore to reduce thread contention
This commit significantly reduces contention on highly concurrent execution of build operations. It moves the coordination to a semaphore.
When a build operation requires a lease, it will first check if it has no children. If it does, it will try to acquire a lease from the
semaphore, and if not, it is queued and put on hold. Then when a semaphore permit is released, we take the first element of the queue and
wake it up (this thread only). This significantly reduces contention as a single thread is waked up at a time, only when it can potentially
acquire a permit (there are still risks to be awaken and not get a permit, but the cost is cheap compared to the previous locking mechanism).

In other words, locking is moved down to the parent build operation, which will be released on demand.