Adding the ability for ivy repo's to use a non-standard pattern
With the current implementation of the PluginRepositories it's not
possible for the user to be able to specify a repo that doens't meet the
"ivy" standard layout. Many people do not, so it's impossible to use
this feature.

This chang will allow for users to specify more info about their
repositories.