Fixed play plugin duplicate artifact error (#1857)
1. Fix for duplicate artifact error "Could not copy MANIFEST.MF / Multiple entries with same key".
After building Gradle with this fix, we're able to build "dist" for the small repro project.
The repro project is attached to the issue #1857.
2. I hope you can wrap it with all the necessary integ test / unit test coverage as needed
because it would be much easier for someone in Gradle team to do it :)
I hope it's fair to give you the fix that works plus clean repro project!
3. Regarding the change from immutable Guava type to standard HashMap:
the field is internal and lightly used so keeping it formally immutable does not bring a lot of value.
This particular Guava type is not optimized for speed but for immutability.
The artifact "rename" value is a exactly the same (we debugged to inspect).
Therefore, it is safe to just replace the value if the key is present.
Simple solution that works :)