Make task execution plan faster
The method that determines the next task was unnecessarily expensive.
This was always the case, but got more pressing as we introduced
the new state lock abstraction, which lead to calling this method
much more often.

We were creating a new worker lease on every iteration, even though
in most cases it would go unused, e.g. because no task is ready or
because all projects are locked. Creating a worker lease is currently
an expensive operation that leads to a lot of garbage. Instead, we
now create one child lease for each task worker and reuse that.

The exit condition for the method was also very expensive, iterating
over the whole execution plan on every iteration. It now only iterates
over the remaining tasks instead, which will usually yield a yes/no
decision after checking very few elements.

Since we now use --max-workers task workers even when --parallel is not
active, we have several threads fighting over the single build model lock.
Instead of iterating over all tasks, only to fail on the model lock check
each time, we check whether there is any project lock available at all.
Since the number of projects is generally much smaller than the number
of tasks, this saves a lot of time.