Use Path instead of ProjectIdentifier for project registry
The use of `ProjectIdentifier` was problematic, in that we were hanging
onto all `DefaultProjectDescriptor` references for the life of the build.
In addition, ProjectIdentifier is not particularly useful as an 'identifier' type.

The registry now contains the identity path for every project in a build,
and can provide a `ProjectComponentIdentifier` instance for each of
these paths.

gradle/composite-builds#37