Fix incremental compiler not detecting removal of a jar on classpath
This commit fixes #1913 by making the incremental compiler rely on the previous compilation jar snapshot
to detect whether a jar has been removed or not. The rationale for this change is that the incremental task
inputs detection has a bug when the collection is ordered: while it will properly detect that something has
changed, the information it returns is incorrect: it would notify of modified jars for jars which haven't
changed, notify of removal jars which have not been removed, and eventually miss the real removal.

A proper fix for incremental task inputs change detection will be issued later. In the meantime, we will
use the information the incremental compiler already knows as the basis for analysis.