Fix deadlock in `DefaultDaemonScanInfo`
This commit fixes a deadlock encountered from time to time on CI. The reason for the deadlock is that we're trying to remove
the same listener from 2 different notifications, one of them being the listener being removed itself. To avoid this, we just
tell the other listener that we're taking care of it by setting the reference to null.