Don't use "smart" lookup method in `DefaultServiceRegistry`
This method was an attempt to workaround problems accessing protected methods, but the generated lookup will
always be `DefaultServiceRegistry` and won't grant access to protected methods of subclasses of the registry.
As a consequence, method handles are only used on public accessible methods. It's worth noting that we should
probably make the contract for providers stronger, since calling `setAccessible(true)` will not be possible
anymore in JDK 9.