Add Isolatable<T> interface
The idea behind this code was to attempt to endown the `ValueSnapshot`
hierarchy with the ability to "isolate" the objects they are
snapshotting.

This commit cannot be merged as-is because it breaks several important
integration tests. The crux of the issue is that
`SerializedValueSnapshot` instances do not contain enough information
to recreate the orginal objects. What's more, I'm not sure they can
be enhanced to capture enough information.