Added a public service `ObjectFactory` that can be used to create objects of various kinds for use in the Gradle model. For now, the only kind of object that can be created is immutable instances of `Named` and it sub-types. These are intended to be used for variant attribute values.
This service is available for injection into plugins and tasks (and anywhere else service injection is available). There is also a `Project.getObjects()` method for access.

Other kinds of construction methods can be added to this service later.

Also added a `Project.getProviders()` method for access to a `ProviderFactory`, so we don't have to smear all of its methods onto `Project`.