Change the “SnapshotTaskInputsOperationDetails” to return the task instance instead of just the path.
Gradle can and will execute two tasks with the same path and identity path. This happens when a task has been declared as a dependency of another task, and then is replaced with another instance. This happens more frequently than you would expect, as many plugins out in the wile overwrite tasks (e.g. shadow plugin).

Therefore, for build scans, the task path is not enough to uniquely identify the task. The only reliable thing is the task object.