Add failing tests for ad hoc tasks not sharing cached results
When two tasks share all their inputs, and have actions from the same classloaders, we allow them to share cached results. However, this is wrong when the actions are actually different. This can happen quite easily with ad hoc tasks, leading to incorrect results being reused. This commit includes a test case demonstrating this that is currently failing.

+review REVIEW-6534