Remove locking from output queue
We were using a LinkedBlockingQueue with a limit of Integer.MAX_VALUE
to buffer messages to the client. This means that we were effectively
never blocking and might as well use a non-blocking queue
implementation instead. The benefit of a non-blocking queue is that
it can be lock-free, thus reducing contention and speeding up the build.