Polish Javadoc of DirectoryBuildCache
In order to have the default size show up in the DSL docs I needed
to add the size as text. This is since we cannot parse the
constant expression written in the source code in our tooling - and
since 5 GB looks much nicer than 5120 MB.
Since the constant then was unused I just inlined it.