Recompile everything whenever jars are reordered
This bugfix makes sure that we recompile everything whenever order of classpath changed, or mutation of the classpath was detected. The reason
is that the order on classpath matters, because a class can be provided in different jars, with a different ABI. For now, we
don't track precisely, when a class is compiled, _from which jar_ is used to be compiled against. This means that if the order
changes, we cannot determine precisely what needs to be recompiled. Considering that re-ordering, or additions/removals on
classpath should not be frequent (updates are still incremental), we recompile everything.