Use BuildOperationListener for buildScan custom values
`BuildOperationNotificationListener` will not work since only one
notification listener is allowed there. And this one is already used by
the build scan plugin.

Note that I removed the application of the build scan custom values
script from `buildSrc` since the listener is build tree scoped and so
the `buildSrc` listener would be present in the main build as well
whereas the build scan plugin already determined that the build is
finished causing `The buildScan extension cannot be used after the build has finished.`
warnings in the main build.