Increase memory for the daemon
The Gradle build needs more than 1G for some more demanding tasks
like a parallel `quickCheck`.

This change also reduces the client VM memory to something reasonable.
It was set to 2G before because we were running the build in the client
VM on CI. Instead we will now be forking single-use daemons on CI, which
only adds ~500ms per build and interoperates better with the memory being
defined in gradle.properties.