Expose the task up to date messages to scans (#2146)
This change makes the existing “execute task” build operation a “rich” operation (i.e. visible to scans) and uses it to convey the up to date messages.

Other task result data that was on TaskStateInternal for the sake of scans has been moved to the result of this op. 

As part of this change, the scan plugin no longer uses TaskStateInternal, TaskExecutionListener or InternalTaskExecutionListener. The latter existed just for build scans, so it has now been removed.