Ignore artifact when relocation exists in pom
This fixes https://github.com/gradle/gradle/issues/1789

When <relocation> element exists in pom, both original artifact
(if it exists) and relocated artifact will be downloaded, which causes
package conflict. For example, org.apache.commons:commons-io:1.3.2 is
relocated to commons-io:commons-io:1.3.2, if a project depends on
org.apache.commons:commons-io:1.3.2 and commons-io:commons-io:2.4, the
two artifacts will both exist in resolved classpath.
(See MavenPomPackagingResolveIntegrationTest.groovy provided in this
commit) This fix checks if relocation exists in pom and ignore its
artifact if necessary. This problem is first introduced in Gradle 2.0.

An integration test is provide to justify the fix.

In addition, an unnecessary artifact retrieval in test and a redundant
System.out.println("") is removed.