Don't use a delegating task action to wait for included build completion
Introduces an `IncludedBuildTaskResource` type that is implemented by the task
that is resolved for an `IncludedBuildTaskReference`. When a task of this type
is encountered in the task graph, it uses a special `TaskInfo` implementation
that does not execute the task, and relies on the `isComplete()` method on the
task instance itself to determine if it has completed.

This removes the execution of the generated task that delegates to the included
build execution, making composite builds more seamless and allowing better
coordination between the consuming and producing builds.