Add support for 'console' output type of CodeNarc plugin.
This fix https://github.com/gradle/gradle/issues/1481 . Although it is
documented that CodeNarc plugin supports 'console' output type, that
type is never implemented. This commit add 'console'
TaskGeneratedSingleFile to CodeNarcReportsImpl, and adjust the options
with which Ant codenarc task is called. To redirect Ant output to Gradle
standard output in 'console' type, the lifecycleLogLevel of AntLogAdapter
is changed to INFO.

Then, a unit test and an integration test are provided.