Include each of the dependency graph resolution failures in `ArtifactCollection.failures` when in lenient mode, rather than grouping them together as causes of a single `ResolveException`.
For artifact/file query methods that are not lenient, fail eagerly when the dependency graph cannot be fully resolved, rather than continuing and attempting to resolve the artifacts. This was the previous behaviour of these methods.

Also removed a duplicate from the exception chain when these methods do fail due to a dependency graph problem.