Disable test for @Destoys with composite builds
The `clean build` and `build clean` use case no longer work
with composite builds. Previously they worked a bit by
accident, as there is no real communication between the
consuming and producing builds to ensure that a task output
is not removed while it is still being used.

This is gradle/composite-builds#114