Pre-populate the repository for Play! tests
This commit introduces a new task, `integTestPrepare`, which is executed before any other Play! integration test
and will populate the Gradle cache found in the integration test home directory with dependencies used by the
integration tests. This should prevent some flakiness in those tests, which tend to timeout because they all try
to download the Scala compiler and its family. While it's in general not an issue, the problem is that the
collection of artifacts to be download is quite large, so cross-process locking may reach the 60s timeout. To
prevent this from happening, we forcefully execute a dummy test during the `integTestPrepare` execution, which
will trigger downloading of dependencies. This test is alone, and will be executed serially before the integration
tests run concurrently in `parallelIntegTest`.