Add test that demonstrates multiple executions of included build
Tasks for an included build are preemptively scheduled when they
are discovered while building the task graph. However certain dependencies
(such as compileOnly) may be discovered after the build has commenced.

In these cases, the included build will be invoked again, resulting in
multiple executions of the same included build. This test demonstrates
this scenario.