Ensure exact task name is supplied for included build task reference
We were incidentally permitting the shorthand notation when creating an
included build task reference. With recent changes, this no longer works
because the task requested doesn't match the task executed.
This commit removes the ability to provide a partially-matched task
name for an included build task reference.

This also fixes an integration test that accidentally relied on partial-name
matching, thus exposing the changed behaviour.