Introduce a `quickFeedbackCrossVersionIntegTest` build type
This build type is similar to `crossVersionIntegTest`, but only runs a subset of the released Gradle versions.
This is done so that we can use it in "commit" builds, for faster feedback, while still running the whole
test suite in a later stage, executed less often (because less likely to fail).

The current selection process is to select the first and last released versions for each major release of
Gradle. There's an option to ban specific versions of Gradle from this list, in which case the selection
process is the same, but without considering the banned versions.