Adds a check for null values in signing properties
If a property such as `signing.keyId` was set to a null value, perhaps by
setting it to the return of `System.getenv('KEY_ID')` when `KEY_ID` is unset,
the PgpSignatoryFactory would throw an NPE when reading those properties while
constructing the PgpSignatory.

This change adds a null check that also bypasses signatory creation when
`required` is false.

Fixes #2267