Fix synchronization in `HttpClientHelper`
This commit fixes the synchronization in `HttpClientHelper` which prevented parallel downloads from happening in case
of an authenticated repository. To fix it, the helper now maintains a map of thread to http context, as recommended
by the HttpClient team. We don't use the JDK thread-local map because we need to clear the contexts once we close
the helper, in order to avoid leaking memory.