Hotfix deadlock in parallel test execution
For each test task we create `maxParallelForks` actors
which receive test classes in a round-robin fashion. These
actors then acquire a worker lease and send the tests to the
actual test process for execution. The problem is that these
actors have a limited queue, so if one of them fails to acquire
a worker lease, then the whole test task comes to a halt, because
the actor blocks the round-robin loop. This can then deadlock
the whole build if there are more test tasks than maximum workers.

This hotfix for Gradle 4.0 just raises the queue size of the actors,
so they won't block the round-robin loop for any realistically-sized project.
The real fix would be to acquire a worker lease *before* we create a new
actor. This will be a follup-up change for 4.1.