Fix conditional logic bug
The idea of the if statement in the code is that we should only be
setting the `-sourcepath` argument when there is sourcepath specified
in the `compileOptions`,
(i.e. `sourcepath != null && !sourcepath.isEmpty()`) or when we are
okay setting an explicitly empty `-sourcepath`
argument (i.e. `!noEmptySourcePath`)

Before this change, if a user called `noEmptySourcepath()` then we
would not be setting the sourcepath even if it was not empty.