Deprecate incubating sourcepath from ComileOptions
The sourcepath property of CompileOptions really should not ever be
necessary for Java compilation. It's also of dubious value for any
compilation task given our ability to model the complete set of
sources in the Gradle DSL.

There is a performance optimiztion used by the `ApiGroovyCompiler`
which is an exception to the above statement about not setting the
sourcepath, but that code can be updated to set the sourcepath
command-line option directly in the `CompileOptions.compilerArgs`
array and we can use some other mechanism to communicate between
the `ApiGroovyCompiler` and the `Compiler<JavaCompileSpec>` used
by the `ApiGroovyCompiler` to communicate that in this case it
makes sense to respect the `sourcepath` argument.