Test binary compatibility on each commit
We add a test which checks if the public API of the current version is
binary compatible to the last released version.
Currently, this test lives in `distributions` until it has found a
better home. We could also run this test as part of sanity check since
it is pretty fast (6s).
The public API definition has been extracted to the root build script.
It could be nice to have the definition in a place where the test
could easily load it without going through using system properties.