Update to ASM6 alpha
This commit bumps the ASM library used in Gradle from 5.1 to 6.0_ALPHA. While being an alpha version, this ASM version is the only
one compatible with the Java 9 class format. This allows us to remove the `Java9ClassReader`, and also makes it possible to re-enable
some tests which were failing due to mocking bugs under JDK 9.