Revert sourcepath deprecation and use custom JavaFileManager (#2356)
* Add the SourcepathIgnoringJavaFileManager

This will allow us to run with an explicitly empty source path, and
not have to worry that Java 9 modules won't build because of it.

* Revert "Deprecate incubating sourcepath from ComileOptions (#2330)"

This reverts commit 33881a3eefd6e334ab83ec308bcbc7c7f3ab5ae8.

It turns out, it actually makes sense to support users' specifying a
sourcepath, but we need to handle it better than we currently do.

* Fix the condition for ignoring the sourcepath

* Add an integration test for Java 9 Jigsaw module compilation

Note: This won't actually run, because Spock seems to not play nicely
with the current Java 9 JDK release (174). I think things will begin
working again once there is a release that has
`--permit-illegal-access` on by default. But, I do think the test is
written correctly.