Merge sourcepaths from compilerArgs with sourcepath property (#2362)
* Merge sourcepaths from compilerArgs with sourcepath property

This change also deprecates specifying a sourcepath in the
compilerArgs. We don't want users doing that so that we can correctly
cache their tasks. But, we also don't want to break them as they
migrate from passing the sourcepath command line argument via the
property on CompileOptions.

* Allow running from non-java 9 JDK

Gradle can be running on an earlier JDK, but have a Java 9 compatible
JDK installed and the test can still be run using that other JDK's
JAVA_HOME.

* Override list method to ignore sources on the classpath

When pretending not to have a sourcepath, it's also important not to
look on the classpath for source files.

* Address review feedback

Specifically:
  - Make a function static instead because it is a pure function.
  - Add a check to avoid a possible `NoSuchElementException`