Configure all projects eagerly when requesting models
The configuration was done in a buildFinished hook until now, which
meant that any buildFinished-calls that were discovered as part of
that configuration were not being called.

The configuration is now done right after loading the projects,
so that all buildFinished hooks are discovered long before that
event is fired.