Remove build-on-demand for composite builds
This changes means that dependencies resolved for buildscript
classpath, and dependencies resolved at configuration time,
will be correctly substituted in a composite build but will not
be built correctly.

We need to fix this for buildscript dependencies.
The new behaviour is intended for configuration-time dependencies
(similar to regular project dependencies resolved at configuration-time)