Disable two Javadoc integration tests for JDK 9
There are problems with multiline javadoc header arguments in JDK 9,
and that version of javadoc, also doesn't allow using the
`@deprecated` javadoc tag, unless the thing being documented is also
annotated with the `@Deprecated` annotation. But, and here's the
kicker, if we go ahead and add `@Deprecated` the `CustomTaglet` code
spews an `IndexOutOfBounds` exception just on JDK 7 on Windows. So,
we can't make everyone happy in this same test, so, we'll just turn it
off in JDK 9 for now.