Don't include module info into shaded jars
This commit fixes the generated Gradle API (and TestKit) jars, so that they don't include
potential `module-info.class` files from bundled dependencies. This was the case since the
introduction of ASM 6, which comes with a module info file.

This is not a general solution, as, theorically, we should care about module info files.
But, the fact is that if we bundle them as is, they just don't match what we are building:
the API jar would be an "ASM" module, instead of a "Gradle API" one.