Add an option to FindBugs to disable analysis progress (#2181)
This fixes https://github.com/gradle/gradle/issues/1646,
https://github.com/gradle/gradle/issues/1307,
https://github.com/gradle/gradle/issues/1756

It has been reported that FindBugs plugin will always print analysis
progress to standard output no matter what the configuration is. It is
because '-progress' command line option is always activated when
staring FindBugs in FindBugsSpecBuilder. This commit add an option to
FindBugs plugin named 'showProgress' to control it. If set to false,
'-progress' will not be passed to FindBugs.