Change sourcepath ignoring logic to a java.lang.reflect.Proxy
There is a strange check in the Java 9 compiler logic which will only
attempt to use the `--release` flag if the `JavaFileManager` passed to
the compiler is a `StandardJavaFileManager`. The prior implementation
broke that expectation. Using reflection, we are able to dynamically
intercept the invocations we care about on the system-provided
`StandardJavaFileManager` without otherwise changing the interface
of the class.

This change  also reenables two tests which were disabled because
Gradle isn't handling relative paths correctly on Java 9. It's still a
mystery why those relative paths are not resolving correctly, but
these tests are more concrened with correct sourcepath merging, so it
should be fine to switch them to using absolute paths.