Update documentation for build cache
Mention that we are now mirroring remote caches locally, and that build cache errors do not fail the build.

The remaining known issues are also removed in `buildCache.adoc`, as they both refer to a missing timeout feature that we are not aware of anyone actually missing as of now.

+review REVIEW-6554