Fix too early configuration of projects in TAPI
https://github.com/gradle/gradle/pull/2397 lead to a regression, as
it changed the order in which the rootProject {} hook is executed
relative to project configuration. This lead the build scan plugin to
fail, as it was now applied to an already configured project and thus
missing most of the configuration events.

The new solution simply disables configure-on-demand when the tooling
client requested any models. This fixes the problem that #2397 was trying
to address in a more elegant way. Since model builders require full configuration
of the project anyway, this does not affect performance.